requirements.txt
torch
numpy
pytorch-ignite